added trading component, and results on local and wandb

+++ train_stocks.py
@@ -10,7 +10,7 @@ from download_data import process_data
 from stock_trading_env import StockTradingEnv
 from finrl.finrl_meta.preprocessor.yahoodownloader import YahooDownloader
 from finrl.finrl_meta.preprocessor.preprocessors import data_split
-
+from trade_stocks import trade_on_test_df
 import os
 
 import warnings
@@ -20,31 +20,35 @@ ENV_NAME = 'stock_trading'
 
 # DATA
 
-# df = download(TRAIN_START_DATE, VALID_END_DATE, ticker_list=config_tickers.DOW_30_TICKER)
-# if not os.path.exists(TRAIN_CSV_NAME):
-df = YahooDownloader(start_date = TRAIN_START_DATE,
-                    end_date = VALID_END_DATE,
-                    ticker_list = config_tickers.DOW_30_TICKER).fetch_data()
-print(f'Data Downloaded! shape:{df.shape}')
+# df = download(TRAIN_START_DATE, TRADE_END_DATE, ticker_list=config_tickers.DOW_30_TICKER)
+if not (os.path.exists(TRAIN_CSV_NAME) or os.path.exists(TEST_CSV_NAME) or os.path.exists(TRADE_CSV_NAME)):
+    df = YahooDownloader(start_date = TRAIN_START_DATE,
+                        end_date = TRADE_END_DATE,
+                        ticker_list = config_tickers.DOW_30_TICKER).fetch_data()
+    print(f'Data Downloaded! shape:{df.shape}')
 
-df_processed = process_data(df, use_technical_indicator=True, technical_indicator_list=INDICATORS, 
-                            use_vix=True, use_turbulence=True, user_defined_feature=False)
+    df_processed = process_data(df, use_technical_indicator=True, technical_indicator_list=INDICATORS, 
+                                use_vix=True, use_turbulence=True, user_defined_feature=False)
 
-print(f'Data Processed! shape:{df_processed.shape}')
+    print(f'Data Processed! shape:{df_processed.shape}')
 
-train = data_split(df_processed, TRAIN_START_DATE, TRAIN_END_DATE)
-trade = data_split(df_processed, VALID_START_DATE, VALID_END_DATE)
+    train = data_split(df_processed, TRAIN_START_DATE, TRAIN_END_DATE)
+    test = data_split(df_processed, TEST_START_DATE, TEST_END_DATE)
+    trade = data_split(df_processed, TRADE_START_DATE, TRADE_END_DATE)
 
-print(f'Train shape: {train.shape} Trade shape: {trade.shape}')
-print('Saving csvs...')
-df.to_csv(ORIGINAL_CSV_NAME)
-df_processed.to_csv(PROCESSED_CSV_NAME)
-train.to_csv(TRAIN_CSV_NAME)
-trade.to_csv(TRADE_CSV_NAME)
-# else:
-#     train = pd.read_csv(TRAIN_CSV_NAME)
-#     trade = pd.read_csv(TRADE_CSV_NAME)
-#     print(f'Train shape: {train.shape}\nTrade shape: {trade.shape}')
+    print(f'Train shape: {train.shape} Test shape: {test.shape} Trade shape: {trade.shape}')
+    print('Saving csvs...')
+    df.to_csv(ORIGINAL_CSV_NAME)
+    df_processed.to_csv(PROCESSED_CSV_NAME)
+    train.to_csv(TRAIN_CSV_NAME)
+    test.to_csv(TEST_CSV_NAME)
+    trade.to_csv(TRADE_CSV_NAME)
+else:
+    print('reading csvs...')
+    train = pd.read_csv(TRAIN_CSV_NAME, index_col='Unnamed: 0')
+    trade = pd.read_csv(TRADE_CSV_NAME, index_col='Unnamed: 0')
+    test = pd.read_csv(TEST_CSV_NAME, index_col='Unnamed: 0')
+    print(f'Train shape: {train.shape} Test shape: {test.shape} Trade shape: {trade.shape}')
 
 # env
 stock_dimension = len(train.tic.unique())
@@ -67,7 +71,7 @@ env_kwargs = {
 }
 
 env = StockTradingEnv(df = train, **env_kwargs)
-
+# print('Asset: ', env.asset_memory)
 # log
 w_config = dict(
   alpha = ACTOR_LR,
@@ -82,8 +86,8 @@ w_config = dict(
   env = ENV_NAME,
   train_start_date = TRAIN_START_DATE,
   train_end_date = TRAIN_END_DATE,
-  trade_start_date = VALID_START_DATE,
-  trade_end_date = VALID_END_DATE,
+  trade_start_date = TRADE_START_DATE,
+  trade_end_date = TRADE_END_DATE,
   hmax = env_kwargs['hmax'],
   initial_amount = env_kwargs['initial_amount'],
   num_stock_shares = env_kwargs['num_stock_shares'],
@@ -103,7 +107,7 @@ w_config = dict(
 PROJECT_NAME = f"pytorch_ddpg_{ENV_NAME.lower()}"
 
 if USE_WANDB:
-    run = wandb.init(project=PROJECT_NAME, tags=["DDPG", "RL"], config=w_config, job_type='train_model')
+    run = wandb.init(project=PROJECT_NAME, tags=["DDPG", "RL"], config=w_config, job_type='train_model') #, resume=RESUME_LAST_WANDB_RUN)
 
 
 
@@ -118,46 +122,104 @@ if not TRAIN_FROM_SCRATCH:
  
 np.random.seed(SEED)
 
+day_counter_total = 0
 score_history = []
 for i in range(starting_episode, TOTAL_EPISODES):
     obs = env.reset()
     done = False
     score = 0
     step_count = 0
-    actor_loss_per_step_list = []
-    critic_loss_per_step_list = []
+    actor_loss_per_episode = 0
+    critic_loss_per_episode = 0
     agent.episode = i
+    cumulative_rewards_per_step_this_episode = []
     while not done:
         act = agent.choose_action(obs)
         new_state, reward, done, info = env.step(act)
         agent.remember(obs, act, reward, new_state, int(done))
         agent.learn()
-        step_count += 1 
+        step_count += 1
+        cumulative_reward = (env.asset_memory[-1] - env_kwargs['initial_amount']) / env_kwargs['initial_amount']
+        cumulative_rewards_per_step_this_episode.append(cumulative_reward)
+        if USE_WANDB:
+            run.log({'Cumulative returns': cumulative_reward, 'days':day_counter_total})
+        day_counter_total += 1
         # run.log({'Actor Loss - STEP': agent.actor_loss_step})
         # run.log({'Critic Loss - STEP': agent.critic_loss_step}) 
-        actor_loss_per_step_list.append(agent.actor_loss_step)
-        critic_loss_per_step_list.append(agent.critic_loss_step)
-        if step_count % 50 == 0:
-            print(f'actor loss step {step_count}: {agent.actor_loss_step}')
-            print(f'critic loss step {step_count}: {agent.critic_loss_step}')
+        actor_loss_per_episode += agent.actor_loss_step
+        critic_loss_per_episode += agent.critic_loss_step
+        # if step_count % 50 == 0:
+        #     print(f'actor loss step {step_count}: {agent.actor_loss_step}')
+        #     print(f'critic loss step {step_count}: {agent.critic_loss_step}')
         # this is a temporal diff method: we learn at each timestep, 
         # unline monte carlo methods where learning is done at the end of an episode
         score += reward
         obs = new_state
         # env.render()
+    
     score_history.append(score)
     if USE_WANDB:
         run.log({'steps per episode': step_count, 'episode': i})
         run.log({'reward': score, 'episode': i})
         # run.log({'reward avg 100 games': np.mean(score_history[-100:]), 'episode': i})
-        run.log({'Actor loss': np.mean(actor_loss_per_step_list), 'episode': i})
-        run.log({'Critic loss': np.mean(critic_loss_per_step_list), 'episode': i})
+        run.log({'Actor loss': actor_loss_per_episode, 'episode': i})
+        run.log({'Critic loss': critic_loss_per_episode, 'episode': i})
 
-    if i % SAVE_CKP_AFTER_EVERY_NUM_EPISODES == 0:
+    
+    if i % SAVE_CKP_AFTER_EVERY_NUM_EPISODES == 0 or i==TOTAL_EPISODES-1:
        agent.save_models()
-    print('='*50)
-    print('episode ', i, 'reward %.2f' % score)
-    print('='*50)
+    # print('='*50)
+    # print('episode ', i, 'reward %.2f' % score)
+    # print('='*50)
+
+    # test cumulative returns on test set
+    
+
+
+    if i % SAVE_REWARD_TABLE_AFTER_EVERY_NUM_EPISODES == 0 or i==TOTAL_EPISODES-1:
+        df_account_value, df_actions, cumulative_rewards_test = trade_on_test_df(df=test, model=agent, train_df=train, env_kwargs=env_kwargs)
+        print('results table....')
+        print(df_account_value.head())
+        results_dir = './results'
+        account_value_csv_name = f'account_value_test_episode_{i}.csv'
+        actions_csv_name = f'daily_actions_test_episode_{i}.csv'
+        results_table_name = f'return_comparison_episode_{i}.csv'
+        df_account_value.to_csv(os.path.join(results_dir, account_value_csv_name))
+        df_actions.to_csv(os.path.join(results_dir, actions_csv_name))
+
+        df = pd.DataFrame(data=[cumulative_rewards_test[-1], max(cumulative_rewards_test), min(cumulative_rewards_test)],
+                        columns=[f'test [{TEST_START_DATE}_{TEST_END_DATE}]'],
+                        index=['Cumulative Return', 'Max Cumulative return', 'Min cumulative return'])
+
+        df[f'train [{TRAIN_START_DATE}_{TRAIN_END_DATE}]'] = [cumulative_rewards_per_step_this_episode[-1], 
+                                                            max(cumulative_rewards_per_step_this_episode),
+                                                            min(cumulative_rewards_per_step_this_episode)]
+        df.to_csv(os.path.join(results_dir, results_table_name))
+
+        if USE_WANDB:
+            df.reset_index(inplace=True)
+            res_table = wandb.Table(dataframe=df) 
+            run.log({f'Cumulative returns Episode {i}': res_table})
+            
+    
+
+
 
 filename = f'{ENV_NAME}-alpha{str(ACTOR_LR)}-beta{str(CRITIC_LR)}batch_{BATCH_SIZE}-400-300.png'
-plotLearning(score_history, filename, window=1)
+plotLearning(score_history, filename, window=10)
+
+# TRADING
+'''
+insample_risk_indicator = train.drop_duplicates(subset=['date'])
+turb_threshold = insample_risk_indicator.turbulence.quantile(0.996)
+turb_threshold = int(round(turb_threshold))
+print('='*100)
+print('='*100)
+print(f'Using Turbulence threshld: {turb_threshold}')
+print('='*100)
+print('='*100)
+
+# trade_env = StockTradingEnv(df=trade, turbulence_threshold=turb_threshold, risk_indicator_col='turbulence', **env_kwargs)
+
+# df_account_value, df_actions = DRLAgent.DRL_prediction(model=agent, environment=trade_env)
+'''
